feat(webpack): update postcss to v8
