+++ packages/webpack/package.json
@@ -12,11 +12,12 @@
     "@nuxt/babel-preset-app": "2.15.6",
     "@nuxt/friendly-errors-webpack-plugin": "^2.5.1",
     "@nuxt/utils": "2.15.6",
+    "autoprefixer": "^10.3.1",
     "babel-loader": "^8.2.2",
     "cache-loader": "^4.1.0",
     "caniuse-lite": "^1.0.30001249",
     "consola": "^2.15.3",
-    "css-loader": "^4.3.0",
+    "css-loader": "^5.2.7",
     "cssnano": "^4.1.11",
     "eventsource-polyfill": "^0.9.6",
     "extract-css-chunks-webpack-plugin": "^4.9.0",
@@ -30,12 +31,11 @@
     "optimize-css-assets-webpack-plugin": "^5.0.8",
     "pify": "^5.0.0",
     "pnp-webpack-plugin": "^1.7.0",
-    "postcss": "^7.0.32",
-    "postcss-import": "^12.0.1",
+    "postcss": "^8.3.6",
+    "postcss-import": "^14.0.2",
     "postcss-import-resolver": "^2.0.0",
-    "postcss-loader": "^3.0.0",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-url": "^8.0.0",
+    "postcss-loader": "^4.3.0",
+    "postcss-url": "^10.1.3",
     "semver": "^7.3.5",
     "std-env": "^2.3.0",
     "style-resources-loader": "^1.4.1",
